requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.18.0
flake8==7.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.9
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mccabe==0.7.0
mongomock==4.3.0
mypy-extensions==1.0.0
narwhals==1.35.0
nodeenv==1.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
protobuf==5.29.4
pyarrow==19.0.1
pycodestyle==2.13.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyflakes==3.3.2
pymongo==4.11.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
sentinels==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
streamlit==1.44.1
streamlit-drawable-canvas==0.9.3
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
uvicorn==0.34.0
virtualenv==20.29.3
watchdog==6.0.0
